@@ -24,12 +24,17 @@
         </resources>
         <plugins>
             <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+                <version>2.2.0.RELEASE</version>
+            </plugin>
+            <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>2.3.2</version>
                 <configuration>
-                    <source>7</source>
-                    <target>7</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                 </configuration>
             </plugin>
             <plugin>
@@ -81,8 +86,8 @@
                         </goals>
                         <configuration>
                             <files>
-                                <file>conf/default.properties</file>
-                                <file>conf/info.properties</file>
+                                <file>${basedir}/conf/default.properties</file>
+                                <file>${basedir}/conf/info.properties</file>
                             </files>
                         </configuration>
                     </execution>
